+++ internal/repository/postgresql/migrations/000006_alter_feeds_tables_add_fulltextsearch_columns.down.sql
@@ -11,4 +11,5 @@ DROP INDEX IF EXISTS idx_feed_entries_fulltextsearch_tsv;
 
 ALTER TABLE feed_entries
 DROP COLUMN summary,
+DROP COLUMN textrank_terms,
 DROP COLUMN fulltextsearch_tsv;