+++ internal/repository/postgresql/migrations/000006_alter_feeds_tables_add_fulltextsearch_columns.up.sql
@@ -16,6 +16,7 @@ SET fulltextsearch_tsv = to_tsvector(replace(title, '.', ' ') || ' ' || replace(
 -- Feed Entries
 ALTER TABLE feed_entries
 ADD COLUMN summary            TEXT NOT NULL DEFAULT '',
+ADD COLUMN textrank_terms     TEXT[],
 ADD COLUMN fulltextsearch_tsv TSVECTOR;
 
 CREATE INDEX idx_feed_entries_fulltextsearch_tsv
@@ -23,4 +24,4 @@ ON feed_entries
 USING gin(fulltextsearch_tsv);
 
 UPDATE feed_entries
-SET fulltextsearch_tsv = to_tsvector(title);
+SET fulltextsearch_tsv = to_tsvector(title) || to_tsvector(array_to_string(textrank_terms, ' '));