+++ docs/src/changelog.md
@@ -12,6 +12,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 #### Feeds
 - For each entry in the list, display the title of the corresponding feed
 - Save and display feed descriptions
+- Extract keywords (significant terms) from entry content/description with TextRank
 - Add full-text search based on feed and entry metadata
 
 ### Changed